Pathtracking

Pathtracking routines

Currently, there are kinds of pathtracking routines implemented

HomotopyContinuation.AffinePredictorCorrector — Type.

AffinePredictorCorrector

This algorithm uses as an prediction step an explicit Euler method. Therefore the prediciton step looks like

\[x_{k+1} = x_k - ΔtJ_H(x, t)^{-1}\frac{∂H}{∂t}(x, t)\]

and the correction step looks like

\[x_{k+1} = x_{k+1} - J_H(x, t)^{-1}H(x, t)\]

This algorithm tracks the path in the affine space.

source

HomotopyContinuation.SphericalPredictorCorrector — Type.

SphericalPredictorCorrector

This algorithm uses as an prediction step an explicit Euler method. For the prediction and correction step the Jacobian is augmented by Hermitian transposed $x^*$ to get a square system. Therefore the prediciton step looks like

\[x_{k+1} = x_k - Δt\begin{bmatrix} J_H(x, t) \\ x^* \end{bmatrix}^{-1} \frac{∂H}{∂t}(x, t)\]

and the correction step looks like

\[x_{k+1} = x_{k+1} - \begin{bmatrix} J_H(x, t) \\ x^* \end{bmatrix}^{-1} H(x, t)\]

After each prediciton and correction the algorithm normalizes $x$ again, i.e. $x$ is then a point on the sphere again.

This algorithm tracks the path in the projective space.

source