Solving homotopies

Solving homotopies

The solve function

The solve function solves homotopies with given starting values.

HomotopyContinuation.solve — Function.

    solve(H::AbstractHomotopy, startvalues_s, [algorithm]; kwargs...)

Solve the homotopy H via homotopy continuation with the given startvalues_s and the given algorithm.

kwargs are the keyword arguments for the solver options.

The default pathtracking algorithm is SphericalPredictorCorrector().

To specify another pathracking algorithm, e.g.AffinePredictorCorrector(), write

    solve(H::AbstractHomotopy, startvalues_s, AffinePredictorCorrector(); kwargs...)

The function also takes polynomials as inputs:

    solve(f::Vector{<:MP.AbstractPolynomial{T}})

solves the polynomial system f via a totaldegree homotopy of type StraightLineHomotopy and the SphericalPredictorCorrector pathtracking routine.

To specify homotopy and pathtracker, use

    solve(f::Vector{<:MP.AbstractPolynomial{T}}, [homotopy], [algorithm]; kwargs...)

Default is homotopy = StraightLineHomotopy and algorithm = SphericalPredictorCorrector. For instance,

    solve(f, GeodesicOnTheSphere, AffinePredictorCorrector())

solves $f=0$ with a GeodesicOnTheSphere homotopy and the AffinePredictorCorrector pathtracking routine.

source

The output of solve is an array of type HomotopyContinuation.Result.

Solver options

The solutions function

The solution function helps to extract information from HomotopyContinuation.Result arrays

HomotopyContinuation.solutions — Function.

solutions(r::Result; success=true, at_infnity=true, only_real=false, singular=true)

Filters the solutions which satisfy the constraints.

source

The result array

The HomotopyContinuation.PathResult struct carries the following informations.

HomotopyContinuation.PathResult — Type.

PathResult(startvalue, pathtracker_result, endgamer_result, solver)

Construct a PathResult for a given startvalue. pathtracker_result is the PathtrackerResult until the endgame radius is reached. endgamer_result is the EndgamerResult resulting from the corresponding endgame.

A PathResult contains:

returncode: One of :success, :at_infinity or any error code from the EndgamerResult

solution::Vector{T}: The solution vector. If the algorithm computed in projective space

and the solution is at infinity then the projective solution is given. Otherwise an affine solution is given if the startvalue was affine and a projective solution is given if the startvalue was projective.

residual::Float64: The value of the infinity norm of H(solution, 0).

newton_residual: The value of the 2-norm of $J_H(\text{solution})^{-1}H(\text{solution}, 0)$

log10_condition_number: A high condition number indicates singularty. See Homotopy.κ for details. The value is the logarithmic condition number (with base 10).

windingnumber: The estimated winding number

angle_to_infinity: The angle to infinity is the angle of the solution to the hyperplane where the homogenizing coordinate is $0$.

real_solution: Indicates whether the solution is real given the defined tolerance at_infinity_tol (from the solver options).

startvalue: The startvalue of the path

iterations: The number of iterations the pathtracker needed.

endgame_iterations: The number of steps in the geometric series the endgamer did.

npredictions: The number of predictions the endgamer did.

predictions: The predictions of the endgamer.

source