Solving homotopies

Solving homotopies

The solve function

The solve function solves homotopies with given starting values.

HomotopyContinuation.solve — Function.

solve(H::AbstractHomotopy, startvalues_s, [algorithm]; kwargs...)

Solve the homotopy H via homotopy continuation with the given startvalues_s and the given algorithm.

solve(f::Vector{<:MP.AbstractPolynomial{T}}; homotopytype=GeodesicOnTheSphere, algorithm=SphericalPredictorCorrector(), kwargs...)
solve(f::MP.AbstractPolynomial{T}; homotopytype=GeodesicOnTheSphere, algorithm=SphericalPredictorCorrector(), kwargs...)

Solves the polynomial system f via homotopy continuation. This uses a totaldegree homotopy of the given homotopytype and uses the given algorithm for pathtracking.

source

The output of solve is an array of type HomotopyContinuation.Result.

The solutions function

The solution function helps to extract information from HomotopyContinuation.Result arrays

HomotopyContinuation.solutions — Function.

solutions(r::Result; success=true, at_infnity=true, only_real=false, singular=true)

Filters the solutions which satisfy the constraints.

source

The result array

The HomotopyContinuation.PathResult struct carries the following informations.

returncode:

solution: the zero that is computed (here it is $[-i,1]$).

singular: boolean that shows whether the zero is singular.

residual: the computed value of $|f([-i,1])|$.

newton_residual:

log10_condition_number:

windingnumber

angle_to_infinity: the algorithms homogenizes the system $f$ and then computes all solutions in projective space. The angle to infinity is the angle of the solution to the hyperplane where the homogenizing coordinate is $0$.

real_solution: boolean that shows whether the zero is real.

startvalue:

iterations:

endgame_iterations:

npredictions: