Solving homotopies

Solving homotopies

The solve function

The solve function solves homotopies with given starting values.

HomotopyContinuation.solve — Function.

    solve(H::AbstractHomotopy, startvalues_s, [algorithm]; kwargs...)

Solve the homotopy H via homotopy continuation with the given startvalues_s and the given algorithm.

    solve(f::Vector{<:MP.AbstractPolynomial{T}})

Solves the polynomial system f via homotopy continuation. This uses a totaldegree homotopy of type StraightLineHomotopy and the SphericalPredictorCorrector pathtracking routine. To specify homotopy and pathtracker, use

solve(f::Vector{<:MP.AbstractPolynomial{T}}, [homotopy], [algorithm])

Default is homotopy = StraightLineHomotopy and algorithm = SphericalPredictorCorrector. For instance, use

solve(f, StraightLineHomotopy, AffinePredictorCorrector())

for solving $f$ with a GeodesicOnTheSphere homotopy and the AffinePredictorCorrector pathtracking routine.

source

The output of solve is an array of type HomotopyContinuation.Result.

Solver options

The solutions function

The solution function helps to extract information from HomotopyContinuation.Result arrays

HomotopyContinuation.solutions — Function.

solutions(r::Result; success=true, at_infnity=true, only_real=false, singular=true)

Filters the solutions which satisfy the constraints.

source

The result array

The HomotopyContinuation.PathResult struct carries the following informations.

PathResult